[MOD] 修改

+++ api/sns/controller/FollowController.php
@@ -24,7 +24,7 @@ class FollowController extends RestUserBaseController
      * PUT  关注/取消关注话题
      */
     public function topic() {
-        if ($this->request->isPut()) {
+        if ($this->request->isPost()) {
             $params = $this->request->param();
             $action = empty($params['action']) ? 'follow' : $params['action'];
 
@@ -33,37 +33,20 @@ class FollowController extends RestUserBaseController
                 $this->error($validate->getError());
             }
 
-            try {
-                $findFollow = SnsTopicFollowModel::withTrashed()
-                    ->where('topic_id', $params['id'])
-                    ->where('user_id', $this->userId)
-                    ->find();
+            $findFollow = SnsTopicFollowModel::where([
+                'topic_id'  => $params['id'],
+                'user_id'   => $this->userId
+            ])->find();
 
-                if ($action == 'follow') {
-                    // 判断是否已经存在关注项了
-                    if (!empty($findFollow)) {
-                        $findFollow->delete_at = 0;
-                        $findFollow->save();
-                    } else {
-                        SnsTopicFollowModel::create([
-                            'topic_id'   =>  $params['id'],
-                            'user_id'    =>  $this->userId
-                        ]);
-                        $topic = SnsTopicModel::get($params['id']);
-                        $topic->follow_count = ['inc', 1];
-                        $topic->save();
-                    }
-                    $this->success('关注成功');
-                } else if ($action == 'cancel') {
-                    if (!empty($findFollow)) {
-                        $findFollow->delete();
-                        $this->success('取消成功');
-                    } else {
-                        $this->error('请先关注');
-                    }
-                }
-            } catch (\Exception $e) {
-                $this->error('系统错误');
+            if ($action == 'follow' && empty($findFollow)) {
+                SnsTopicFollowModel::create([
+                    'topic_id'   =>  $params['id'],
+                    'user_id'    =>  $this->userId
+                ]);
+                $this->success('关注成功');
+            } else if ($action == 'cancel' && !empty($findFollow)) {
+                $findFollow->delete();
+                $this->success('取消成功');
             }
 
         }
@@ -73,7 +56,7 @@ class FollowController extends RestUserBaseController
      * 关注/取消关注帖子
      */
     public function post() {
-        if ($this->request->isPut()) {
+        if ($this->request->isPost()) {
             $params = $this->request->param();
             $action = empty($params['action']) ? 'follow' : $params['action'];
 
@@ -82,34 +65,20 @@ class FollowController extends RestUserBaseController
                 $this->error($validate->getError());
             }
 
-            try {
-                $findFollow = SnsPostFollowModel::withTrashed()
-                    ->where('post_id', $params['id'])
-                    ->where('user_id', $this->userId)
-                    ->find();
+            $findFollow = SnsPostFollowModel::where([
+                'post_id'  => $params['id'],
+                'user_id'   => $this->userId
+            ])->find();
 
-                if ($action == 'follow') {
-                    // 判断是否已经存在关注项了
-                    if (!empty($findFollow)) {
-                        $findFollow->delete_at = 0;
-                        $findFollow->save();
-                    } else {
-                        SnsPostFollowModel::create([
-                            'post_id'   =>  $params['id'],
-                            'user_id'    =>  $this->userId
-                        ]);
-                    }
-                    $this->success('关注成功');
-                } else if ($action == 'cancel') {
-                    if (!empty($findFollow)) {
-                        $findFollow->delete();
-                        $this->success('取消成功');
-                    } else {
-                        $this->error('请先关注');
-                    }
-                }
-            } catch (\Exception $e) {
-                $this->error('系统错误');
+            if ($action == 'follow' && empty($findFollow)) {
+                SnsPostFollowModel::create([
+                    'post_id'   =>  $params['id'],
+                    'user_id'    =>  $this->userId
+                ]);
+                $this->success('关注成功');
+            } else if ($action == 'cancel' && !empty($findFollow)) {
+                $findFollow->delete();
+                $this->success('取消成功');
             }
 
         }
@@ -119,7 +88,7 @@ class FollowController extends RestUserBaseController
      * 关注/取消关注用户
      */
     public function user() {
-        if ($this->request->isPut()) {
+        if ($this->request->isPost()) {
             $params = $this->request->param();
             $action = empty($params['action']) ? 'follow' : $params['action'];
 
@@ -128,42 +97,54 @@ class FollowController extends RestUserBaseController
                 $this->error($validate->getError());
             }
 
-            try {
-                $findFollow = SnsUserFollowModel::withTrashed()
-                    ->where('follow_id', $params['id'])
-                    ->where('user_id', $this->userId)
-                    ->find();
+            $findFollow = SnsUserFollowModel::where([
+                'follow_uid'  => $params['id'],
+                'user_id'   => $this->userId
+            ])->find();
 
-                if ($action == 'follow') {
-                    // 判断是否已经存在关注项了
-                    if (!empty($findFollow)) {
-                        $findFollow->delete_at = 0;
-                        $findFollow->save();
-                    } else {
-                        SnsUserFollowModel::create([
-                            'follow_id'   =>  $params['id'],
-                            'user_id'    =>  $this->userId
-                        ]);
-                        // 建立被关注的消息通知
-                        SnsNoticeFansModel::create([
-                            'notice_uid'    =>  $params['id'],
-                            'fans_id'       =>  $this->userId,
-                            'content'       =>  "{$this->user['user_nickname']}关注了你"
-                        ]);
-                    }
-                    $this->success('关注成功');
-                } else if ($action == 'cancel') {
-                    if (!empty($findFollow)) {
-                        $findFollow->delete();
-                        $this->success('取消成功');
-                    } else {
-                        $this->error('请先关注');
-                    }
-                }
-            } catch (\Exception $e) {
-                $this->error('系统错误');
+            if ($action == 'follow' && empty($findFollow)) {
+                SnsUserFollowModel::create([
+                    'follow_uid'   =>  $params['id'],
+                    'user_id'    =>  $this->userId
+                ]);
+
+                // 建立被关注的消息通知
+                SnsNoticeFansModel::create([
+                    'notice_uid'    =>  $params['id'],
+                    'fans_id'       =>  $this->userId,
+                    'content'       =>  "{$this->user['user_nickname']}关注了你"
+                ]);
+
+                $this->success('关注成功');
+            } else if ($action == 'cancel' && !empty($findFollow)) {
+                $findFollow->delete();
+                $this->success('取消成功');
             }
+        }
+    }
 
+    /**
+     * 获取所有我关注的话题id列表
+     */
+    public function topicList() {
+        if ($this->request->isGet()) {
+            $data = SnsTopicFollowModel::where('user_id', $this->userId)
+                ->select();
+            $this->success('成功', $data);
+        }
+    }
+
+    /**
+     * 获取所有我关注的用户id列表
+     * @throws \think\db\exception\DataNotFoundException
+     * @throws \think\db\exception\ModelNotFoundException
+     * @throws \think\exception\DbException
+     */
+    public function userList() {
+        if ($this->request->isGet()) {
+            $data = SnsUserFollowModel::where('user_id', $this->userId)
+                ->select();
+            $this->success('成功', $data);
         }
     }
